UserLoginService Json serialization

+++ src/main/java/pl/sda/intermediate/weather/model/Wind.java
@@ -15,6 +15,6 @@ public class Wind {
     public Double speed;
     @SerializedName("deg")
     @Expose
-    public Integer deg;
+    public Double deg;
 
 }
